Time: 474 ms (54.89%) | Memory: 0B (100.00%) - LeetSync

+++ 1161-project-employees-i/project-employees-i.sql
@@ -1,6 +1,6 @@
 SELECT 
   p.project_id, 
-  ROUND(AVG(e.experience_years), 2) AS average_years
+  ROUND(AVG(e.experience_years), 2)::NUMERIC AS average_years
 FROM Project p
 INNER JOIN Employee e
 USING (employee_id)
